Converted to 16 bitwidth

+++ examples/hed/compute_segmentation.py
@@ -61,7 +61,7 @@ def main(caffe_mode):
 
 		img_number_str = test_lst[idx][(test_lst[idx].rfind('/') + 1):] # e.g. "0.0.png"
 		print 'fuse.shape ==', fuse.shape
-		png.from_array(fuse, 'L').save('fuse_output_' + img_number_str)
+		png.from_array(fuse.astype(np.uint16), 'L').save('fuse_output_' + img_number_str)
 
 if __name__ == '__main__':
 	parser = argparse.ArgumentParser(description='Test out hed + save some images!')
